% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/competitiveness_data_annual.R
\name{build_annual_competitiveness}
\alias{build_annual_competitiveness}
\title{Rebuild the competitiveness data set}
\usage{
build_annual_competitiveness(
  download_data_www = FALSE,
  year_start = 1994,
  year_end = 2020
)
}
\description{
Function used to create the data set.
}
