% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{competitiveness_data_macro}
\alias{competitiveness_data_macro}
\title{Macroeconomic data on competitiveness}
\format{
A data frame with XX rows and XX variables:
\describe{
  \item{iso3c}{The iso3c code}
  \item{year}{The year of observation}
  ...
}
}
\source{
\url{http://www.spatial-competition.com/}
}
\usage{
competitiveness_data_macro
}
\description{
Data on competitiveness in the EU
}
\keyword{datasets}
