% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_oecd_functions.R
\name{get_oecd_av_wages}
\alias{get_oecd_av_wages}
\title{OECD data on average wages}
\usage{
get_oecd_av_wages(download_data, countries_considered, first_year, last_year)
}
\description{
Gets data on average wages from OECD dataset 'AV_AN_WAGE'
}
\details{
Gets data on average wages from OECD dataset 'AV_AN_WAGE'. Only download
 average wages in PPP dollar ('USDPPP').
}
