% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_oecd_functions.R
\name{get_oecd_data}
\alias{get_oecd_data}
\title{Gets OECD data}
\usage{
get_oecd_data(download_data, countries_considered, first_year, last_year)
}
\arguments{
\item{download_data}{If TRUE, data will be downloaded from the internet}

\item{countries_considered}{Vector of iso3c codes of the countries for
which data should be assembled.}

\item{first_year}{First year for which data is to be collected (numeric).}

\item{last_year}{Last year for which data is to be collected (numeric).}
}
\description{
Calls all functions to collect and merge OECD data
}
\seealso{
Other download_helpers: 
\code{\link{get_worldbank}()},
\code{\link{setup_swiid}()}
}
\concept{download_helpers}
