% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_oecd_functions.R
\name{get_oecd_debt_data}
\alias{get_oecd_debt_data}
\title{Get OECD debt data}
\usage{
get_oecd_debt_data(download_data, countries_considered, first_year, last_year)
}
\description{
Get variables from OECD dataset 'FIN_IND_FBS' on debt.
}
\details{
Access to the OECD debt dataset 'FIN_IND_FBS'. From this dataset, the
 following five variables are retrieved: "DBTS1GDP", "DBTS11GDP",
 "DBTS12GDP", "DBTS13GDP", "DBTS14_S15GDI".
}
