% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_oecd_functions.R
\name{get_oecd_finance_data}
\alias{get_oecd_finance_data}
\title{Get OECD finance data}
\usage{
get_oecd_finance_data(
  download_data,
  countries_considered,
  first_year,
  last_year
)
}
\description{
Get data on finance from OECD dataset 'MEI_FIN'.
}
\details{
Retrieves the following finance-related variables from the OECD dataset
 'MEI_FIN': IRLT in its annual version.
}
