% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_oecd_functions.R
\name{get_oecd_pub_debt_data}
\alias{get_oecd_pub_debt_data}
\title{Get OECD public debt data}
\usage{
get_oecd_pub_debt_data(
  download_data,
  countries_considered,
  first_year,
  last_year
)
}
\description{
Gets data on public debt from OECD dataset 'NAAG'.
}
