% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{load_rda}
\alias{load_rda}
\title{Loads an object from an rda file}
\usage{
load_rda(file, object_name = NULL)
}
\arguments{
\item{file}{Path to the `.rda` file}

\item{object_name}{The name of the object to be returned; can be NULL}
}
\value{
If `object_name` is not NULL the object; if yes, a list of all
 objects in the `file`
}
\description{
Loads an object from an rda file
}
