% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rci_data_annual}
\alias{rci_data_annual}
\title{Regional data on competitiveness}
\format{
A data frame with XX rows and XX variables:
\describe{
  \item{iso3c}{The iso3c code}
  \item{year}{The year of observation}
  ...
}
}
\source{
\url{https://ec.europa.eu/regional_policy/en/information/maps/regional_competitiveness/}
}
\usage{
rci_data_annual
}
\description{
Data on regional competitiveness in the EU
}
\keyword{datasets}
