% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CSR_database.R
\name{setup_imp_scores}
\alias{setup_imp_scores}
\title{Setup the implementation scores from the comission}
\usage{
setup_imp_scores()
}
\description{
Takes the csv file that was created from the excel file obtained from the
 EU webpage and provides a data.table with the mean implementation scores
 and means for the MIP, SGP, INTEGRATED and SGP/MIT related country-specific
 recommendations.
}
