% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcc-indicators.R
\name{setup_pcc}
\alias{setup_pcc}
\title{Setup data from the quaterly report on price and cost competitiveness}
\usage{
setup_pcc(download_data)
}
\description{
The data is a relative index with 2010=100. The function only considers
 annual data, quaterly and monthly data is not considered.
 The data contains information about the relative position of the countries.
 The comparison groups are the following:
 EA19: Belgium, Germany, Estonia, Greece, Spain, France, Ireland, Italy,
     Cyprus, Latvia, Lithuania, Luxembourg, Netherlands, Malta, Austria,
     Portugal, Slovenia, Slovakia, Finland.
 EU27: EA19 +
     Bulgaria, Czech Republic, Denmark, Croatia, Hungary, Poland,
     Romania, Sweden.
 IC37: EU27 +
     Australia, Canada, Japan, Mexico, New Zealand, Norway, Switzerland,
     Turkey, United Kingdom, USA
 More information about the variables can be obtained via the labels. The
 latter can be accessed via `View()` or `Hmisc::label()`.
 The country codes are iso3c codes except the two aggregated regions
 'Euro_Area' and 'European_Union', corresponding to groups EA19 and EU27.
}
