% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_solt_functions.R
\name{setup_swiid}
\alias{setup_swiid}
\title{Get data from SWIID 9 database}
\usage{
setup_swiid(
  download_data,
  file_name,
  countries_considered,
  first_year,
  last_year
)
}
\arguments{
\item{download_data}{If TRUE, data will be downloaded from the internet}

\item{file_name}{The name of the `.rda` file containing the data.}

\item{countries_considered}{Vector of iso3c codes of the countries for
which data should be assembled.}

\item{first_year}{First year for which data is to be collected (numeric).}

\item{last_year}{Last year for which data is to be collected (numeric).}
}
\description{
Gets pre and post tax Ginis from the standardized world income inequality
 database provided by Solt. Since names of data files change whenever the
 data is updated one must download the raw data manualle. The `.rda` file
 from the zip should be used as an input to this functino.
}
\seealso{
Other download_helpers: 
\code{\link{get_oecd_data}()},
\code{\link{get_worldbank}()}
}
\concept{download_helpers}
